support nvidia/NVIDIA-Nemotron-3-Super-120B-A12B-NVFP4 -tp1

+++ python/tokenspeed/runtime/layers/logits_processor.py
@@ -471,7 +471,7 @@ class LogitsProcessor(nn.Module):
                     .view(logits.size(0), -1)
                 )
 
-        logits = logits[:, : self.config.vocab_size].contiguous()
+        logits = logits[:, : self.config.vocab_size].float().contiguous()
 
         if self.final_logit_softcapping:
             fused_softcap_generic(logits, self.final_logit_softcapping)
